refactor: Update .utils import path
