+++ kennard_stone/utils/__init__.py
@@ -1,0 +1,6 @@
+from ._pairwise import is_installed, pairwise_distances
+
+__all__ = (
+    "is_installed",
+    "pairwise_distances",
+)
